# docs/spec/api-v1-map.openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "qc_flags": {
            "items": {
              "$ref": "#/components/schemas/QcFlagResponse"
            },
            "title": "Qc Flags",
            "type": "array"
          },
          "qc_rule_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qc Rule Version"
          },
          "qc_status": {
            "enum": [
              "raw",
              "qc_passed",
              "qc_suspect",
              "qc_failed",
              "missing",
              "qc_unchecked"
            ],
            "title": "Qc Status",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "id",
          "station_id",
          "timestamp",
          "parameter",
          "source",
          "qc_status",
          "qc_flags"
        ],
        "title": "ObservationResponse",
        "type": "object"
      },
      "PaginatedResponse_ForecastSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ForecastSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[ForecastSummary]",
        "type": "object"
      },
...
      "PaginatedResponse_StationSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[StationSummary]",
        "type": "object"
      },
...
      "StationDetail": {
        "properties": {
          "basin_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basin Id"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "forecast_targets": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Targets"
          },
          "gauging_status": {
            "title": "Gauging Status",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/GeoCoordResponse"
          },
          "measured_parameters": {
            "items": {
              "type": "string"
            },
            "title": "Measured Parameters",
            "type": "array"
          },
          "model_assignments": {
            "items": {
              "$ref": "#/components/schemas/ModelAssignmentResponse"
            },
            "title": "Model Assignments",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "network": {
            "title": "Network",
            "type": "string"
          },
          "no_floor": {
            "title": "No Floor",
            "type": "boolean"
          },
          "ownership": {
            "title": "Ownership",
            "type": "string"
          },
          "regulation_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulation Type"
          },
          "station_kind": {
            "title": "Station Kind",
            "type": "string"
          },
          "station_status": {
            "title": "Station Status",
            "type": "string"
          },
          "thresholds": {
            "items": {
              "$ref": "#/components/schemas/ThresholdResponse"
            },
            "title": "Thresholds",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "weather_sources": {
            "items": {
              "$ref": "#/components/schemas/WeatherSourceResponse"
            },
            "title": "Weather Sources",
            "type": "array"
          },
          "wigos_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wigos Id"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "location",
          "station_kind",
          "station_status",
          "network",
          "ownership",
          "measured_parameters",
          "timezone",
          "gauging_status",
          "created_at",
          "updated_at",
          "no_floor",
          "thresholds",
          "model_assignments",
          "weather_sources"
        ],
        "title": "StationDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The subset of /api/v1 the flow map reads (Plan 402). Not served \u2014 openapi_url stays None on the live app; this is a committed, drift-tested static artefact. Bearer auth is added explicitly below: the live routes read the Authorization header from the raw Request, so FastAPI's own generator sees no security scheme.",
    "title": "SAPPHIRE Flow \u2014 flow map contract",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/stations": {
      "get": {
        "description": "Lists stations. A consumer, reviewer or admin token may call this; a consumer/reviewer token's result is silently filtered to its scoped stations rather than 404ing.",
        "operationId": "list_stations_api_v1_stations_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_StationSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed query value or station_id (error envelope: {\"error\": <message>, \"detail\": null})."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Stations",
        "tags": [
          "api-stations"
        ]
      }
    },
    "/api/v1/stations/{station_id}": {
      "get": {
        "description": "Station detail. Any authenticated token role may call this; an out-of-scope or unknown station is 404. A malformed station_id is unguarded today and returns 500, not 400.",
        "operationId": "get_station_api_v1_stations__station_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "station_id",
            "required": true,
            "schema": {
              "title": "Station Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown or out-of-scope station/forecast \u2014 the same body either way (never reveals which)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Station",
        "tags": [
          "api-stations"
        ]
      }
    },
    "/api/v1/stations/{station_id}/forecasts": {
      "get": {
        "description": "Forecast summaries for a station, with QC status and typed qc_flags. For an activated CHWRR tenant this contains only the current human-selected publication per issue time, including a selected forecast later marked superseded by an automatic retry. Consumer and reviewer scopes are identical; admin is unscoped but sees the same publication gate. An out-of-scope station is 404.",
        "operationId": "list_forecasts_api_v1_stations__station_id__forecasts_get",
        "parameters": [
          {
            "in": "path",
            "name": "station_id",
            "required": true,
            "schema": {
              "title": "Station Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "in": "query",
            "name": "parameter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameter"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "description": "Return only forecasts assessed PARTIAL or DEGRADED input quality (Plan 253 OD-2). Forecasts with no assessment recorded (unknown) are never included.",
            "in": "query",
            "name": "degraded_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "Return only forecasts assessed PARTIAL or DEGRADED input quality (Plan 253 OD-2). Forecasts with no assessment recorded (unknown) are never included.",
              "title": "Degraded Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ForecastSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed query value or station_id (error envelope: {\"error\": <message>, \"detail\": null})."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown or out-of-scope station/forecast \u2014 the same body either way (never reveals which)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Forecasts",
        "tags": [
          "api-stations"
        ]
      }
    },
    "/api/v1/stations/{station_id}/observations": {
      "get": {
        "description": "Observations for a station and parameter window, with QC status and typed qc_flags. Any authenticated token role may call this; an out-of-scope or unknown station is 404.",
        "operationId": "list_observations_api_v1_stations__station_id__observations_get",
        "parameters": [
          {
            "in": "path",
            "name": "station_id",
            "required": true,
            "schema": {
              "title": "Station Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parameter",
            "required": true,
            "schema": {
              "title": "Parameter",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "qc_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Qc Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObservationResponse"
                  },
                  "title": "Response List Observations Api V1 Stations  Station Id  Observations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed query value or station_id (error envelope: {\"error\": <message>, \"detail\": null})."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown or out-of-scope station/forecast \u2014 the same body either way (never reveals which)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Observations",
        "tags": [
          "api-stations"
        ]
      }
    },
...
  }
}